separated parts of index.ts; cleanup
